refresh db seed for user_id, add rte middleware

+++ routes/web.php
@@ -9,36 +9,35 @@ use App\Http\Controllers\ListingController;
 //All Listings
 Route::get('/', [ListingController::class, 'index']);
 
-
-//Show create form
-Route::get('/listings/create', [ListingController::class, 'create']);
+//Show Create form
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 //Store Listing Data 
-Route::post('/listings', [ListingController::class, 'store']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
 //Show edit form 
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 //Update Listing
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 //Delete Listing
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
 //View Single Listing 
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
 //Show Register create form
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
 //Create new User
 Route::post('/users', [UserController::class, 'store']);
 
 //Log User out
-Route::post('/logout', [UserController::class, 'logout']);
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
 
 //Show Login Form
-Route::get('/login', [UserController::class, 'login']);
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
 
 //Login User
 Route::post('/users/authenticate', [UserController::class, 'authenticate']);
